Moved render_page_links() function to separate file and added require_once's for that file
